# openapi.yaml
openapi: 3.0.3
info:
  title: Preference Syncs API
  version: >
    1.0.0
  description: >
    API for managing applications, users, and preference synchronisation.
...
paths:
...
  /app:
    post:
      operationId: CreateApp
      tags:
        - Apps
      summary: >
        Create an app
      description: >
        Creates a new app owned by the authenticated user. Requires an authenticated admin user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppRequest'
      responses:
        '200':
          description: >
            Application created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppResponse'
        default:
          description: >
            Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /app/{app}:
    get:
      operationId: GetApp
      tags:
        - Apps
      summary: >
        Get an app
      description: >
        Retrieves the details of an app by its ID. Requires an authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - name: app
          in: path
          required: true
          description: >
            The unique identifier of the app
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: >
            Application retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppResponse'
        default:
          description: >
            Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: ModifyApp
      tags:
        - Apps
      summary: >
        Modify an app
      description: >
        Updates the name, description, or limits of an existing app. Requires an authenticated admin user.
      security:
        - bearerAuth: []
      parameters:
        - name: app
          in: path
          required: true
          description: >
            The unique identifier of the app
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppRequest'
      responses:
        '200':
          description: >
            Application modified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppResponse'
        default:
          description: >
            Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: DeleteApp
      tags:
        - Apps
      summary: >
        Delete an app
      description: >
        Permanently deletes an app by its ID. Requires an authenticated admin user.
      security:
        - bearerAuth: []
      parameters:
        - name: app
          in: path
          required: true
          description: >
            The unique identifier of the app
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: >
            Application deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        default:
          description: >
            Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Bearer token authentication
  schemas:
    AppRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          description: >
            Name of the application
        desc:
          type: string
          nullable: true
          description: >
            Description of the application
        max_payload:
          type: integer
          format: int64
          nullable: true
          description: >
            Maximum payload size in bytes
        max_user:
          type: integer
          format: int64
          nullable: true
          description: >
            Maximum number of users
      description: >
        Payload for creating or updating an application
    AppResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: >
            Unique identifier of the application
        name:
          type: string
          description: >
            Name of the application
        desc:
          type: string
          description: >
            Description of the application
        max_payload:
          type: integer
          format: int64
          description: >
            Maximum payload size in bytes
        max_user:
          type: integer
          format: int64
          description: >
            Maximum number of users
        created_by:
          type: string
          format: uuid
          description: >
            Unique identifier of the user who created the application
      description: >
        Application response payload
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: >
            Indicates whether the operation was successful
      description: >
        Generic success response
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          description: >
            Indicates whether an error occurred
        message:
          type: string
          description: >
            Error message description
      description: >
        Standard error response