# server/apps.go
package server

import (
	"encoding/json/v2"
	"net/http"
	"uuid"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/render"
	"github.com/samber/oops"

	"github.com/khwong-c/pref-syncs/features"
	"github.com/khwong-c/pref-syncs/server/middlewares"
)

type appReqPayload struct {
	Name       *string `json:"name"`
	Desc       *string `json:"desc"`
	MaxPayload *int    `json:"max_payload"`
	MaxUser    *int    `json:"max_user"`
}

type appRspPayload struct {
	ID         uuid.UUID `json:"id"`
	Name       string    `json:"name"`
	Desc       string    `json:"desc"`
	MaxPayload int       `json:"max_payload"`
	MaxUser    int       `json:"max_user"`
	CreatedBy  uuid.UUID `json:"created_by"`
}

func (s *Server) HandleNewApp(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	uid := s.auth.GetUserID(ctx)
	oops.FromContext(ctx).Assert(uid != uuid.Nil())

	payload := &appReqPayload{}
	if err := json.UnmarshalRead(r.Body, payload); err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Invalid Payload", http.StatusBadRequest,
		)
		return
	}

	newApp, err := s.appLogic.CreateApp(ctx, &features.AppUpsertReq{
		CreatedBy:  uid,
		Desc:       payload.Desc,
		Name:       payload.Name,
		MaxPayload: payload.MaxPayload,
		MaxUser:    payload.MaxUser,
	})
	if err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Failed to create App", 0,
		)
		return
	}
	render.JSON(w, r, appRspPayload{
		ID:         newApp.ID,
		Name:       newApp.Name,
		Desc:       newApp.Desc,
		MaxPayload: newApp.MaxPayload,
		MaxUser:    newApp.MaxUser,
		CreatedBy:  newApp.CreatedBy,
	})
}

func (s *Server) HandleGetApp(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	id, err := uuid.Parse(chi.URLParam(r, "app"))
	if err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Invalid UUID", http.StatusBadRequest,
		)
		return
	}
	app, err := s.appLogic.GetApp(ctx, id)
	if err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Failed to retrieve app", 0,
		)
		return
	}
	render.JSON(w, r, appRspPayload{
		ID:         app.ID,
		Name:       app.Name,
		Desc:       app.Desc,
		MaxPayload: app.MaxPayload,
		MaxUser:    app.MaxUser,
	})
}

func (s *Server) HandleModifyApp(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	appID, err := uuid.Parse(chi.URLParam(r, "app"))
	if err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Invalid UUID", http.StatusBadRequest,
		)
		return
	}

	userID := s.auth.GetUserID(ctx)

	req := &appReqPayload{}
	if err := json.UnmarshalRead(r.Body, req); err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Invalid Payload", http.StatusBadRequest,
		)
		return
	}

	modApp, err := s.appLogic.ModifyApp(
		ctx, userID, appID,
		&features.AppUpsertReq{
			Name:       req.Name,
			Desc:       req.Desc,
			MaxPayload: req.MaxPayload,
			MaxUser:    req.MaxUser,
		},
	)
	if err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Failed to modify App", 0,
		)
		return
	}
	render.JSON(w, r, appRspPayload{
		ID:         modApp.ID,
		Name:       modApp.Name,
		Desc:       modApp.Desc,
		MaxPayload: modApp.MaxPayload,
		MaxUser:    modApp.MaxUser,
	})
}

func (s *Server) HandleDeleteApp(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	id, err := uuid.Parse(chi.URLParam(r, "app"))
	if err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Invalid UUID", http.StatusBadRequest,
		)
		return
	}
	err = s.appLogic.DeleteApp(ctx, id)
	if err != nil {
		middlewares.SimpleHTTPError(
			ctx, w, err, "Failed to delete app", 0,
		)
		return
	}
	render.JSON(w, r, struct {
		Success bool `json:"success"`
	}{true})
}